Small fix to firmware reload bug when image=0

+++ fem_sw/src/commandProcessor.h
@@ -55,7 +55,7 @@ struct clientStatus {
 
 void commandProcessorThread(void* arg);
 void disconnectClient(struct clientStatus* pState, int *pIndex, fd_set* pFdSet, u8 *pNumConnectedClients);
-void commandHandler(struct protocol_header* pRxHeader, struct protocol_header *pTxHeader, u8* pRxPayload, u8* pTxPayload, u8* pMux, XGpio *pGpio, u8* pReloadRequested);
+void commandHandler(struct protocol_header* pRxHeader, struct protocol_header *pTxHeader, u8* pRxPayload, u8* pTxPayload, u8* pMux, XGpio *pGpio, int* pReloadRequested);
 int socketRead(int sock, u8* pBuffer, unsigned int numBytes);
 int validateHeaderContents(struct protocol_header *pHeader);
 void flushSocket(int sock, void *mem, int len);
